Data model with shared pointers to Entities
Convert player ownership of entities to be via pointers to ids. Each player has a hashmap, with an x,y location as key, and a pointer to an entity as the value.

Cells also now have ownership of entities via a hashmap from player id to entity pointer.

Entities also know the id of their owner.

These changes disrupt some json serial and deserialization. Places that still need to be changed to be logically compatible with new format have been marked TODO. All code is currently compilation compatible.