Increase Windows pipe buffer size
Avoids hanging when writing messages, up to a point. Unlike Unix,
a process that continues to send commands but does not read input
will eventually hang the game environment, as Windows waits to
write when the buffer is full. We could fix this permanently by
using overlapping I/O, but that's a bigger engineering effort
than we have time for right now.

Fixes #289.