Compile Julia bots during compilation phase
Julia uses lazy compilation and compiles code when it is first used. Before this PR bots written in Julia are compiled when the code if first used, i.e during the initialization phase. This unfairly limits the amount of code and packages that can be used in Julia projects.

This PR triggers compilation of Julia projects during the compilation phase of Halite3.